requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Banco de dados
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2_cffi==21.3.0

# Utilidades
pydantic==2.8.2
python-dotenv==1.0.1

# SDK AWS
boto3==1.35.24

# Dados financeiros e agendamento
yfinance==0.2.48
APScheduler==3.10.4
pandas==2.2.3

# Desenvolvimento e qualidade de código
pylint==3.0.3